requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.1
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6